Remove custom logging file.
There is no reason for a custom log file if some exceptions are not
logged to it, as of both versions often report event exceptions with the
main logger, and not the plugin logger. Therefore I'm removing it again
and rather include some information about the host in the server's log.