core: StatusResponse: Apply custom player slots if response version == 9999
If we already applied the custom player slots because the client was
up-to-date and then try to process the same response again, it would
look like the client is outdated. In that case the custom player slots
would be reverted again. Processing a response twice can happen on
Bukkit with Paper + ProtocolLib.

To fix this, assume that the client is up-to-date if response version
== 9999.

Fixes #332.