core: Fix replacements that contain $
The replacement functionality of java.util.regex.Matcher has a rather
unfortunate property that the replacements to appendReplacement(...)
must always have $ chars escaped (since it's used to insert groups).
SLP does not make use of that functionality, so it would just be annoying
and stupid overhead to call Matcher.quoteReplacement(...) for all of them.

Instead, introduce a LiteralMatcherReplacer that works mostly like
the replacement functionality in the Matcher, but without requiring
to escape all the replacements.