Use builder class for creating formatted chat messages
The `ChatBuilder` class encapsulates everything related to chat message formatting into a single class. This class is especially useful when logic is involved in the building of a chat message (like in the `buildCommandHelp` method) because the `StringBuilder` object you would normally use is encapsulated inside of `ChatBuilder`.

Modifying the `ServerCommandSender.sendMessage()` method signature to accept a `CharSequence` instead of a `String` would eliminate the need to call `toString()` at the end of every message. Doing this would still allow you to pass `String` objects into the method because `String` implements `CharSequence`.