Fix NPE thrown during favicon cache creation
The `ServerListPlusCore` constructor takes a reference to a plugin. Inside of this constructor, the plugin's `createFaviconCache` method is called. In the case of the `ServerListPlusServer` plugin, this method assumes that the plugin's `ServerListPlusCore` field is non-null, which was a wrong assumption (the `ServerListPlusCore` object's constructor is still being executed). This caused a null reference to be passed into the `FaviconCache` constructor.

The solution is to pass the plugin object into the `FaviconCache` constructor so that `ServerListPlusPlugin.getCore()` can be called.