Create FaviconCache class to reduce code duplication (#354)
* Create FaviconCache class to reduce code duplication

This class wraps a "LoadingCache" object, which is used by each of the plugins to cache favicons.

As far as I can tell, the "FaviconHelper.loadSafely" method is only used by this favicon caching code, so I integrated the code from this method into the new FaviconCache class and deleted the method.

* Modify method signature to directly return `FaviconCache`

* Create `FaviconCache` once in every plugin

* Convert tab indents to spaces

* Several changes related to `FaviconCache` class

* Set `ServerListPlusCore` field in constructor instead of setter method
* Clear cache by calling `FaviconCache.clear()` instead of passing null into `reloadFaviconCache`
* Rename `ServerListPlusPlugin.reloadFaviconCache` to `createFaviconCache`

See: https://github.com/Minecrell/ServerListPlus/pull/354#discussion_r644863662

* Remove the 1.7.10 compatibility check and revert the `FaviconCache.contains` method back to `getIfPresent`

* Remove unnecessary cache call

* Stop using `Optional.ifPresent`

https://github.com/Minecrell/ServerListPlus/pull/354#discussion_r645043784

* Drop the Guava-to-Java conversion of `Optional`

* Fix NPE thrown during favicon cache creation

The `ServerListPlusCore` constructor takes a reference to a plugin. Inside of this constructor, the plugin's `createFaviconCache` method is called. In the case of the `ServerListPlusServer` plugin, this method assumes that the plugin's `ServerListPlusCore` field is non-null, which was a wrong assumption (the `ServerListPlusCore` object's constructor is still being executed). This caused a null reference to be passed into the `FaviconCache` constructor.

The solution is to pass the plugin object into the `FaviconCache` constructor so that `ServerListPlusPlugin.getCore()` can be called.

* Minor style fixes