Sponge: Try getting/setting protocol version with SpongeAPI
On SpongeAPI 10 the implementation-specific code does not work anymore.
To fix this once and for all try to do it via proper APIs instead.

NOTE: This is not supported in SpongeAPI yet. Until this is merged copy
the modified interfaces to the stub source set to allow compiling.
See https://github.com/SpongePowered/SpongeAPI/pull/2456