Bukkit: Avoid ProtocolLib error on newer MC versions
The handshake packet no longer has the Protocol enum in newer MC versions,
but a new "ClientIntent" enum. There is support for that in the latest
ProtocolLib versions, but not in a release yet.

For now avoid a crash by using readSafely() and just record all logins
if the type of the handshake is unavailable. This will produce additional
load on the request cache, but should be largely negligible.

Fixes #420